+++ src/main/scala/de/htwg/se/twothousandfortyeight/controller/actorInternalDslImpl/Move.scala
@@ -1,6 +1,6 @@
 package de.htwg.se.twothousandfortyeight.controller.actorInternalDslImpl
 
 case class Move(seq: String*) {
-  seq.foreach(e => Player.play(e))
+  seq.foreach(e => MoveActor.move(e))
   sys.exit
 }
